description

+++ src/main/java/com/bookstore/domain/Book.java
@@ -23,7 +23,7 @@ public class Book {
     private String category;
     private int numberOfPages;
     private String format;
-    private int isbn;
+    private Long isbn;
     private double shippingWeight;
     private double listPrice;
     private double ourPrice;
@@ -108,11 +108,11 @@ public class Book {
         this.format = format;
     }
 
-    public int getIsbn() {
+    public Long getIsbn() {
         return isbn;
     }
 
-    public void setIsbn(int isbn) {
+    public void setIsbn(Long isbn) {
         this.isbn = isbn;
     }
